Uploaded Files
Only included the Asset files and build files. You download the apk to
your android device and run it from there to try it out. Otherwise
import the assets to your own unity project if you want to experiment
with it. Be sure to set your build option to Android and set up the
Android SDK before you try using these files in Unity.